Implement tests for creating user category

+++ rest-api-server/src/main/java/br/com/tiagoamp/timetracker/repository/CategoryEntity.java
@@ -1,8 +1,5 @@
 package br.com.tiagoamp.timetracker.repository;
 
-import br.com.tiagoamp.timetracker.model.Category;
-import br.com.tiagoamp.timetracker.model.User;
-
 import javax.persistence.*;
 
 @Entity
@@ -11,7 +8,7 @@ public class CategoryEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -22,20 +19,20 @@ public class CategoryEntity {
 
     public CategoryEntity() { }
 
-    public CategoryEntity(Integer id, String name, String description, UserEntity userEntity) {
+    public CategoryEntity(Long id, String name, String description, UserEntity userEntity) {
         this.id = id;
         this.name = name;
         this.description = description;
         this.user = userEntity;
     }
 
-    public CategoryEntity(Integer id, String name, String description) {
+    public CategoryEntity(Long id, String name, String description) {
         this(id, name, description, null);
     }
 
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) {
+    public Long getId() { return id; }
+    public void setId(Long id) {
         this.id = id;
     }
     public String getName() { return name; }
